refactored creation search logic into a util file
made search query persistent

cleanup
